+++ presentation/home/src/main/java/com/hilingual/presentation/home/component/dialog/DiaryContinueDialog.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hilingual.presentation.home.component
+package com.hilingual.presentation.home.component.dialog
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
